integer number of points.

+++ broadbean/blueprint.py
@@ -759,7 +759,7 @@ def _subelementBuilder(blueprint: BluePrint, SR: int,
     # and raise an exception if the segment ends up with less than
     # two points
 
-    intdurations = np.zeros(len(newdurations))
+    intdurations = np.zeros(len(newdurations), dtype=int)
 
     for ii, dur in enumerate(newdurations):
         int_dur = round(dur*SR)
